+++ internal/quickstart/container.go
@@ -2,45 +2,53 @@ package quickstart
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/charmbracelet/bubbles/key"
+	"github.com/charmbracelet/bubbles/spinner"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
+	"github.com/muesli/reflow/wordwrap"
 
 	"ldcli/internal/flags"
 )
 
-// step is an identifier for each step in the quick-start flow.
-type step int
+const (
+	defaultProjKey = "default"
+	defaultEnvKey  = "test"
+)
 
 const (
-	createFlagStep step = iota
-	chooseSDKStep
-	showSDKInstructionsStep
+	_ = iota
+	stepCreateFlag
+	stepChooseSDK
+	stepShowSDKInstructions
+	stepToggleFlag
 )
 
 // ContainerModel is a high level container model that controls the nested models wher each
 // represents a step in the quick-start flow.
 type ContainerModel struct {
-	currentStep step
-	err         error
-	flagKey     string
-	flagsClient flags.Client
-	quitMsg     string
-	quitting    bool
-	sdk         sdkDetail
-	steps       []tea.Model
+	accessToken  string
+	baseUri      string
+	currentModel tea.Model
+	currentStep  int
+	err          error
+	flagKey      string
+	flagsClient  flags.Client
+	quitting     bool
+	sdk          sdkDetail
+	totalSteps   int
+	width        int
 }
 
-func NewContainerModel(flagsClient flags.Client) tea.Model {
+func NewContainerModel(flagsClient flags.Client, accessToken string, baseUri string) tea.Model {
 	return ContainerModel{
-		currentStep: createFlagStep,
-		flagsClient: flagsClient,
-		steps: []tea.Model{
-			NewCreateFlagModel(flagsClient),
-			NewChooseSDKModel(),
-			NewShowSDKInstructionsModel(),
-		},
+		accessToken:  accessToken,
+		baseUri:      baseUri,
+		currentModel: NewCreateFlagModel(flagsClient, accessToken, baseUri),
+		currentStep:  1,
+		flagsClient:  flagsClient,
+		totalSteps:   4,
 	}
 }
 
@@ -49,124 +57,96 @@ func (m ContainerModel) Init() tea.Cmd {
 }
 
 func (m ContainerModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	var (
-		cmd     tea.Cmd
-		updated tea.Model
-	)
+	var cmd tea.Cmd
 	switch msg := msg.(type) {
+	case tea.WindowSizeMsg:
+		m.width = msg.Width
 	case tea.KeyMsg:
 		switch {
-		case key.Matches(msg, keys.Quit):
+		case key.Matches(msg, pressableKeys.Quit):
 			m.quitting = true
-
-			return m, tea.Quit
-		case key.Matches(msg, keys.Enter):
+			cmd = tea.Quit
+		case key.Matches(msg, pressableKeys.Back):
 			switch m.currentStep {
-			case createFlagStep:
-				updated, cmd := m.steps[createFlagStep].Update(msg)
-				if model, ok := updated.(createFlagModel); ok {
-					if model.err != nil {
-						m.err = model.err
-						if model.quitting {
-							m.quitMsg = model.quitMsg
-							m.quitting = true
-
-							return m, cmd
-						}
-
-						return m, nil
-					}
-
-					m.flagKey = model.flagKey
-					m.currentStep += 1
-				}
-			case chooseSDKStep:
-				updated, _ := m.steps[chooseSDKStep].Update(msg)
-				if model, ok := updated.(chooseSDKModel); ok {
-					m.sdk = model.selectedSDK
-					m.currentStep += 1
-				}
-			case showSDKInstructionsStep:
-				_, cmd := m.steps[showSDKInstructionsStep].Update(msg)
-				m.currentStep += 1
-
-				return m, cmd
-			default:
+			case stepCreateFlag:
+				// can't go back
+			case stepChooseSDK:
+				m.currentStep -= 1
+				m.currentModel = NewCreateFlagModel(m.flagsClient, m.accessToken, m.baseUri)
+			case stepShowSDKInstructions:
+				m.currentStep -= 1
+				m.currentModel = NewChooseSDKModel(m.sdk.index)
+				cmd = m.currentModel.Init()
+			case stepToggleFlag:
+				m.currentStep -= 1
+				m.currentModel = NewShowSDKInstructionsModel(
+					m.accessToken,
+					m.baseUri,
+					m.sdk.canonicalName,
+					m.sdk.displayName,
+					m.sdk.url,
+					m.flagKey,
+				)
+				cmd = m.currentModel.Init()
 			}
 		default:
 			// delegate all other input to the current model
-			updated, cmd := m.steps[m.currentStep].Update(msg)
-			m.steps[m.currentStep] = updated
-
-			return m, cmd
+			m.currentModel, cmd = m.currentModel.Update(msg)
 		}
-		switch m.currentStep {
-		case showSDKInstructionsStep:
-			updated, cmd = m.steps[showSDKInstructionsStep].Update(fetchSDKInstructionsMsg{
-				canonicalName: m.sdk.canonicalName,
-				flagKey:       m.flagKey,
-				name:          m.sdk.displayName,
-			})
-			if model, ok := updated.(showSDKInstructionsModel); ok {
-				model.sdk = m.sdk.displayName
-				m.steps[showSDKInstructionsStep] = model
-			}
-		default:
-		}
-	case errMsg:
-		m.err = msg.err
-	case noInstructionsMsg:
+	case choseSDKMsg:
+		m.currentModel = NewShowSDKInstructionsModel(
+			m.accessToken,
+			m.baseUri,
+			msg.sdk.canonicalName,
+			msg.sdk.displayName,
+			msg.sdk.url,
+			m.flagKey,
+		)
+		cmd = m.currentModel.Init()
+		m.sdk = msg.sdk
 		m.currentStep += 1
-
-		return m, cmd
+	case createdFlagMsg:
+		m.currentModel = NewChooseSDKModel(0)
+		m.flagKey = msg.flagKey // TODO: figure out if we maintain state here or pass in another message
+		m.currentStep += 1
+		m.err = nil
+	case errMsg:
+		m.currentModel, cmd = m.currentModel.Update(msg)
+	case noInstructionsMsg:
+		// skip the ShowSDKInstructionsModel and move along to toggling the flag
+		m.currentModel = NewToggleFlagModel(
+			m.flagsClient,
+			m.accessToken,
+			m.baseUri,
+			m.flagKey,
+			m.sdk.kind,
+		)
+		m.currentStep += 1
+	case fetchedSDKInstructions, fetchedEnv, selectedSDKMsg, toggledFlagMsg, spinner.TickMsg:
+		m.currentModel, cmd = m.currentModel.Update(msg)
+		m.err = nil
+	case showToggleFlagMsg:
+		m.currentModel = NewToggleFlagModel(
+			m.flagsClient,
+			m.accessToken,
+			m.baseUri,
+			m.flagKey,
+			m.sdk.kind,
+		)
+		m.currentStep += 1
 	default:
+		log.Printf("container default: %T\n", msg)
 	}
 
 	return m, cmd
 }
 
 func (m ContainerModel) View() string {
-	// TODO: remove after creating more steps
-	if m.currentStep > showSDKInstructionsStep {
-		return fmt.Sprintf("created flag %s\nselected the %s SDK", m.flagKey, m.sdk.displayName)
+	out := fmt.Sprintf("\nStep %d of %d\n"+m.currentModel.View(), m.currentStep, m.totalSteps)
+
+	if m.quitting {
+		return ""
 	}
 
-	out := fmt.Sprintf("\nStep %d of %d\n"+m.steps[m.currentStep].View(), m.currentStep+1, len(m.steps))
-	if m.err != nil {
-		if m.quitting {
-			out := m.quitMsg + "\n\n"
-			out += m.err.Error()
-
-			return lipgloss.
-				NewStyle().
-				SetString(out).
-				Render() + "\n"
-		}
-
-		// show error and stay on the same step
-		out += "\n" + lipgloss.
-			NewStyle().
-			SetString(m.err.Error()).
-			Render() + "\n"
-
-		return out
-	}
-
-	return out
-}
-
-type keyMap struct {
-	Enter key.Binding
-	Quit  key.Binding
-}
-
-var keys = keyMap{
-	Enter: key.NewBinding(
-		key.WithKeys("enter"),
-		key.WithHelp("enter", "select"),
-	),
-	Quit: key.NewBinding(
-		key.WithKeys("ctrl+c"),
-		key.WithHelp("q", "quit"),
-	),
+	return wordwrap.String(out, m.width)
 }