+++ CHANGELOG.md
@@ -1,5 +1,66 @@
 # Changelog
 
+## [0.5.0](https://github.com/launchdarkly/ldcli/compare/v0.4.0...v0.5.0) (2024-04-05)
+
+
+### Features
+
+* show help text ([#138](https://github.com/launchdarkly/ldcli/issues/138)) ([69d7f5e](https://github.com/launchdarkly/ldcli/commit/69d7f5ee03579fc6c3cb34fb00fb6f07ba90bf84))
+
+## [0.4.0](https://github.com/launchdarkly/ldcli/compare/v0.3.0...v0.4.0) (2024-04-05)
+
+
+### Features
+
+* publish docker image ([#99](https://github.com/launchdarkly/ldcli/issues/99)) ([a294ce0](https://github.com/launchdarkly/ldcli/commit/a294ce063e608a6dbee5ff8a2fb32f329d643083))
+
+
+### Bug Fixes
+
+* create flag prompt should be on the same line as text ([#135](https://github.com/launchdarkly/ldcli/issues/135)) ([52aa92b](https://github.com/launchdarkly/ldcli/commit/52aa92b647710f86f531b69afb3be50baf452a8d))
+
+## [0.3.0](https://github.com/launchdarkly/ldcli/compare/v0.2.0...v0.3.0) (2024-04-04)
+
+
+### Features
+
+* Can go back to choose SDK page from show SDK instructions ([#120](https://github.com/launchdarkly/ldcli/issues/120)) ([6900bc6](https://github.com/launchdarkly/ldcli/commit/6900bc6301bf668830ca029867f6d9de4c555d85))
+
+
+### Bug Fixes
+
+* add instructions to continue from show sdk step ([#117](https://github.com/launchdarkly/ldcli/issues/117)) ([4baf2d0](https://github.com/launchdarkly/ldcli/commit/4baf2d0b2850b3979267f1d2ab964b5a295a27e3))
+* fix the step count ([#121](https://github.com/launchdarkly/ldcli/issues/121)) ([1943c45](https://github.com/launchdarkly/ldcli/commit/1943c4590587f0a3b84e60c6a1628a2ed37254b6))
+* remove id from goreleaser homebrew build config ([#125](https://github.com/launchdarkly/ldcli/issues/125)) ([56538cd](https://github.com/launchdarkly/ldcli/commit/56538cd34bb3ea908e4622a110ab6f486c288b86))
+
+## [0.2.0](https://github.com/launchdarkly/ldcli/compare/v0.1.0...v0.2.0) (2024-04-03)
+
+
+### Features
+
+* add ldcli formula to homebrew-tap ([#108](https://github.com/launchdarkly/ldcli/issues/108)) ([1d638dc](https://github.com/launchdarkly/ldcli/commit/1d638dc0bf23f1e9fab96718a0a0058a74284dc0))
+* Add more help in error message ([#72](https://github.com/launchdarkly/ldcli/issues/72)) ([6221983](https://github.com/launchdarkly/ldcli/commit/62219838f1c815695f6529252ded80170f267dcf))
+* add sdk instructions step to quickstart ([#91](https://github.com/launchdarkly/ldcli/issues/91)) ([bf4aba6](https://github.com/launchdarkly/ldcli/commit/bf4aba61651c7bc359157e39abceb9b4bf7b101e))
+* add toggle flag step ([#111](https://github.com/launchdarkly/ldcli/issues/111)) ([9cd4018](https://github.com/launchdarkly/ldcli/commit/9cd401883f0b2a9b1d84fde095658316091ecd3d))
+* add toggle on and off aliases to update flag ([#82](https://github.com/launchdarkly/ldcli/issues/82)) ([7b6c6f1](https://github.com/launchdarkly/ldcli/commit/7b6c6f1e160598353ab0ed0e93cbc1bd3dd04360))
+* alias command to invite members ([#84](https://github.com/launchdarkly/ldcli/issues/84)) ([7002866](https://github.com/launchdarkly/ldcli/commit/7002866263f510637546c3fea02608083cfb689b))
+* better flag create error handling ([#85](https://github.com/launchdarkly/ldcli/issues/85)) ([708925b](https://github.com/launchdarkly/ldcli/commit/708925b4978f85539fffca513c84434f00b5d08f))
+* Create choose SDK view ([#89](https://github.com/launchdarkly/ldcli/issues/89)) ([7518423](https://github.com/launchdarkly/ldcli/commit/7518423da3712fa2986fe20935428d7336d21b81))
+* create command to get an environment ([#96](https://github.com/launchdarkly/ldcli/issues/96)) ([19b4ede](https://github.com/launchdarkly/ldcli/commit/19b4ede40d836855896c06105e7f0fb01d7e6161))
+* create command to invite members ([#68](https://github.com/launchdarkly/ldcli/issues/68)) ([e1a2ca5](https://github.com/launchdarkly/ldcli/commit/e1a2ca5d09415143cc0673674d83d29fc1a99e17))
+* create prod-ready quickstart command ([#75](https://github.com/launchdarkly/ldcli/issues/75)) ([7768bfa](https://github.com/launchdarkly/ldcli/commit/7768bfae10842292e615c5da8b18e3ee94066049))
+* pass in optional role flag when using members invite ([#90](https://github.com/launchdarkly/ldcli/issues/90)) ([b51470d](https://github.com/launchdarkly/ldcli/commit/b51470d54d6c63fcc478b64770a2c6e91a2539fb))
+* setting version dynamically ([#83](https://github.com/launchdarkly/ldcli/issues/83)) ([7e4e794](https://github.com/launchdarkly/ldcli/commit/7e4e7943a425689e342e34c55a613bf02b7c44c2))
+* Show error along with message and quit if applicable ([#87](https://github.com/launchdarkly/ldcli/issues/87)) ([f076a59](https://github.com/launchdarkly/ldcli/commit/f076a59772b85b095411f63f35e1c66bee0304d0))
+
+
+### Bug Fixes
+
+* fix err messages for toggle flag command ([#112](https://github.com/launchdarkly/ldcli/issues/112)) ([9b385e9](https://github.com/launchdarkly/ldcli/commit/9b385e9a0816ba1677aaf170b1186f86f69d786a))
+* members create cmd should invite multiple members ([#103](https://github.com/launchdarkly/ldcli/issues/103)) ([c575248](https://github.com/launchdarkly/ldcli/commit/c575248c6ab058622dd6e55f2b312656743e4cff))
+* rebind projKey flag to update subcommand ([#62](https://github.com/launchdarkly/ldcli/issues/62)) ([1f0f898](https://github.com/launchdarkly/ldcli/commit/1f0f898511716becdf57ecd90a81ee84c56b1217))
+* remove create flag placeholder text ([#114](https://github.com/launchdarkly/ldcli/issues/114)) ([3e8624d](https://github.com/launchdarkly/ldcli/commit/3e8624d6e70c45285143a5bbded5c917fff829bc))
+
 ## 0.1.0 (2024-03-22)
 
 ### Miscellaneous Chores